refactor: move BatchStock routes to batchStockController
