+++ src/main/java/com/example/mercadofrescos/controller/PurchaseOrderController.java
@@ -18,7 +18,6 @@ import java.util.List;
 public class PurchaseOrderController {
 
     private final IPurchaseOrderService service;
-    private final IBatchStockService serviceBatchStock;
 
     /**
      * Cadastra um lote com o estoque de produtos que o compõe
@@ -57,32 +56,6 @@ public class PurchaseOrderController {
         return ResponseEntity.ok(orderUpdated);
     }
 
-    /**
-     * Ordena pela data de vencimento
-     * @author Ma, Giovanna e Gabriel
-     * @param sectionId e 'days'
-     * @return retorna uma lista de 'batchStocks'
-     */
-    @GetMapping("/due-date")
-    public ResponseEntity<BatchStockResponseDTO> getBatchStockOrderByDueDate(@RequestParam Integer days, @RequestParam Long sectionId) {
-        BatchStockResponseDTO batchStock = serviceBatchStock.getBatchStockOrderByDueDate(days, sectionId);
-        return ResponseEntity.ok(batchStock);
-    }
 
-    /**
-     * Retorna uma lista de lotes no prazo de validade solicitada com uma determinada categoria de produto de forma crescente
-     * @author Ma, Gabriel e Giovanna
-     * @param days, category e orderBy
-     * @return retorna uma lista de 'batchStocks'
-     */
-    @GetMapping("/due-date/list")
-    public ResponseEntity<BatchStockResponseDTO> getBatchStockOrderByDueDateAndCategory(
-            @RequestParam Integer days,
-            @RequestParam String category,
-            @RequestParam(required=false) OrderBy orderBy
-    ) {
-        BatchStockResponseDTO batchStock = serviceBatchStock.getBatchStockOrderByDueDateAndCategory(days, category, orderBy);
-        return ResponseEntity.ok(batchStock);
-    }
 
 }
